Fixed - TExponentialSpin was not emitting valueChanged() after setValue() was called. Fixed - Radio buttons for colour choice in Style window were no acting as a group. Added start of Axes toolbar. Fixed - Style window was still using integer alpha value.

+++ src/gui/style.h
@@ -103,7 +103,7 @@ class StyleWindow : public QWidget
 	// -- Alpha Control
 	void on_ColourAlphaOwnAlphaRadio_clicked(bool checked);
 	void on_ColourAlphaFixedAlphaRadio_clicked(bool checked);
-	void on_ColourAlphaFixedAlphaSpin_valueChanged(int value);
+	void on_ColourAlphaFixedAlphaSpin_valueChanged(double value);
 
 
 	/*
